requirements.txt
GitPython
bottle
